Retrieval_grounded/requirements.txt
# Retrieval-Grounded QA Benchmark
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
tqdm>=4.65.0
psutil>=5.9.0
google-generativeai>=0.8.0
PyYAML>=6.0.0
pytest>=8.0.0
